+++ src/module/settings.ts
@@ -312,12 +312,15 @@ export const registerSettings = function () {
   });
 
   getGame().settings.register(VARIANT_ENCUMBRANCE_MODULE_NAME, 'useEquippedUnequippedItemCollectionFeature', {
-    name: getGame().i18n.localize(VARIANT_ENCUMBRANCE_MODULE_NAME + '.setting.useEquippedUnequippedItemCollectionFeature.name'),
-    hint: getGame().i18n.localize(VARIANT_ENCUMBRANCE_MODULE_NAME + '.setting.useEquippedUnequippedItemCollectionFeature.hint'),
+    name: getGame().i18n.localize(
+      VARIANT_ENCUMBRANCE_MODULE_NAME + '.setting.useEquippedUnequippedItemCollectionFeature.name',
+    ),
+    hint: getGame().i18n.localize(
+      VARIANT_ENCUMBRANCE_MODULE_NAME + '.setting.useEquippedUnequippedItemCollectionFeature.hint',
+    ),
     scope: 'world',
     config: true,
     type: Boolean,
     default: false,
   });
-  
 };